requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
mistralai==1.5.0
google-genai==1.5.0
python-dotenv==1.0.1
jinja2==3.1.4
